Refactor feed-stage to remove excerpt and smart-content fields
Removed the excerpt and smart-content field components and updated feed-stage logic to use a single 'content' field for rendering post content. Adjusted field definitions and view configuration to reflect this change, simplifying the feed display and code structure.
