+++ src/social-web/components/fields/index.ts
@@ -1,13 +1,11 @@
 export * from './avatar';
 export * from './content';
 export * from './date';
-export * from './excerpt';
 export * from './follow-status';
 export * from './metadata';
 export * from './modified';
 export * from './name';
 export * from './object-type';
-export * from './smart-content';
 export * from './status';
 export * from './title';
 export * from './webfinger';
